// src/api/bybit/types.ts
// Generated by https://quicktype.io

export interface WalletBalanceRes {
  ret_code: number;
  ret_msg: string;
  ext_code: string;
  ext_info: string;
  result: WalletCryptoBalance;
  time_now: string;
  rate_limit_status: number;
  rate_limit_reset_ms: number;
  rate_limit: number;
}

export interface WalletCryptoBalance {
  equity: number;
  available_balance: number;
  used_margin: number;
  order_margin: number;
  position_margin: number;
  occ_closing_fee: number;
  occ_funding_fee: number;
  wallet_balance: number;
  realised_pnl: number;
  unrealised_pnl: number;
  cum_realised_pnl: number;
  given_cash: number;
  service_cash: number;
}
export interface OrderOptions {
  symbol: string;
  qty: number;
  side: OrderSide;
  type: OrderType;
  orderLinkId?: string;
}

export enum OrderType {
  limit = 'LIMIT',
  market = 'MARKET',
  limitMaker = 'LIMIT_MAKER',
}

export enum OrderSide {
  buy = 'Buy',
  sell = 'Sell',
}